+++ app/routes/queries.listings.tsx
@@ -171,26 +171,10 @@ const Index = () => {
     setError(undefined)
   }
 
-  // Paywall logic
-  const showPaywall =
-    !loaderData.isLoggedIn || !loaderData.hasPremium || loaderData.needsRefresh
-  const handleLogin = () => {
-    navigate('/discord-login')
-  }
-  const handleSubscribe = () => {
-    window.open(DISCORD_SERVER_URL, '_blank')
-  }
-
   return (
     <PageWrapper>
       <div className="py-3">
-        <PremiumPaywall
-          show={showPaywall}
-          isLoggedIn={loaderData.isLoggedIn}
-          hasPremium={loaderData.hasPremium}
-          needsRefresh={loaderData.needsRefresh}
-          onLogin={handleLogin}
-          onSubscribe={handleSubscribe}>
+        <PremiumPaywall loaderData={loaderData}>
           <SmallFormContainer
             title="Get Item Listing Details"
             onClick={onSubmit}
